Refactored everything - the relative import errors persist no matter what I try for the tests. Will work on packaging in the background, but want to keep a functioning copy before that.
