+++ tests/context.py
@@ -3,16 +3,16 @@ import sys
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
-from bbpow.utxo import UTXO_INPUT, UTXO_OUTPUT
-from bbpow.decoder import Decoder
-from bbpow.block import Block, calc_merkle_root, merkle_proof
-from bbpow.blockchain import Blockchain
-from bbpow.database import DataBase
-from bbpow.formatter import Formatter
-from bbpow.headers import Header
-from bbpow.miner import mine_a_block
-from bbpow.node import Node
-from bbpow.timestamp import utc_to_seconds, utc_timestamp, seconds_to_utc
-from bbpow.wallet import Wallet
-from bbpow.transactions import Transaction, MiningTransaction
-from bbpow.api import create_app, run_app
+from api import create_app, run_app
+from block import Block, calc_merkle_root, merkle_proof
+from blockchain import Blockchain
+from database import DataBase
+from decoder import Decoder
+from formatter import Formatter
+from headers import Header
+from miner import mine_a_block
+from node import Node
+from timestamp import utc_timestamp, seconds_to_utc, utc_to_seconds
+from transactions import MiningTransaction, Transaction
+from utxo import UTXO_INPUT, UTXO_OUTPUT
+from wallet import Wallet
