+++ tests/helpers.py
@@ -9,10 +9,8 @@ from hashlib import sha256
 import logging
 
 from basicblockchains_ecc.elliptic_curve import secp256k1
-
-from bbpow.formatter import Formatter
 from .context import UTXO_INPUT, UTXO_OUTPUT, Transaction, MiningTransaction, utc_to_seconds, Header, Block, Node, \
-    Blockchain, DataBase, Wallet, mine_a_block
+    Blockchain, Wallet, mine_a_block, Formatter
 
 # --- Constants --- #
 curve = secp256k1()
